Working through scroll issues with new get next offers
