+++ Scuti/Scripts/UI/Store/OfferSummaryTallSmallContainer.cs
@@ -13,6 +13,12 @@ namespace Scuti.UI
         public OfferSummaryPresenterContainer Small;
         public OfferSummaryPresenterBase Tall;
 
+        protected override void Awake()
+        {
+            base.Awake();
+            Clear();
+        }
+
         protected void Start()
         {
             Small.InjectContainer(this);
@@ -24,7 +30,7 @@ namespace Scuti.UI
             Small.Clear();
             Tall.Clear();
             Tall.gameObject.SetActive(false);
-            Small.gameObject.SetActive(false);
+            Small.gameObject.SetActive(true);
         }
