--- Scuti/Scripts/UI/Store/OffersPresenterUniversal_old.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 36d61f23019f339499ae8764eebffff5
-timeCreated: 1571151693
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
